refactor: consolidate user tables and fix type errors
- Create shared UserTable component
- Refactor UserManagementTable and RecentUsersTable to use UserTable
- Fix useToastOptions usage in Dashboard
- Fix missing import in HighlightCard
